move hyper1 impl into new client module
