+++ rust-runtime/aws-smithy-http-client/examples/client-ring.rs
@@ -3,10 +3,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-use aws_smithy_http_client::hyper_1::{CryptoMode, HyperClientBuilder};
+use aws_smithy_http_client::{Builder, CryptoMode};
 
 fn main() {
-    let _client = HyperClientBuilder::new()
-        .crypto_mode(CryptoMode::Ring)
-        .build_https();
+    let _client = Builder::new().crypto_mode(CryptoMode::Ring).build_https();
 }
