+++ rust-runtime/aws-smithy-http-client/examples/custom-dns.rs
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-use aws_smithy_http_client::hyper_1::{CryptoMode, HyperClientBuilder};
+use aws_smithy_http_client::{Builder, CryptoMode};
 use aws_smithy_runtime_api::client::dns::{DnsFuture, ResolveDns};
 use std::net::{IpAddr, Ipv4Addr};
 
@@ -17,7 +17,7 @@ impl ResolveDns for StaticResolver {
 }
 
 fn main() {
-    let _client = HyperClientBuilder::new()
+    let _client = Builder::new()
         .crypto_mode(CryptoMode::Ring)
         .build_with_resolver(StaticResolver);
 }
