+++ rust-runtime/aws-smithy-http-client/src/test_util/wire.rs
@@ -333,7 +333,7 @@ impl WireMockServer {
     /// **Note**: This must be used in tandem with [`Self::dns_resolver`]
     pub fn http_client(&self) -> SharedHttpClient {
         let resolver = self.dns_resolver();
-        crate::hyper_1::build_with_fn(None, move || {
+        crate::client::build_with_fn(None, move || {
             hyper_util::client::legacy::connect::HttpConnector::new_with_resolver(
                 resolver.clone().0,
             )
